Add failing key tx test
Change the default key transformation to only allow binaries. This will
make it have the same assumption that previous user defined key transforms
have had since they were created. It will fail if it gets passed a
{tombstone, Key} tuple, which is the problem being fixed here.
Run a merge with completely expired files, one of which contains a
tombstone. This will trigger the path that passes the {tombstone, Key}
instead of just the Key to the key transformer. This will fail with a
function_clause error as expected.