Add failing scan error test
This test demonstrates that when Bitcask's scan files operation fails,
a keydir object is created, not marked ready and left behind. Subsequent
open operations are unable to succeed because they try to wait for this
existing keydir to be ready when its data loads. But this will never
happen, so sadness ensues.