Fix scan error deadlock
If a file scan error happens during a Bitcask open operation, remember
to release the newly created keydir.  Before, that keydir was left
behind. Subsequent open operations would see it, see that it was not
ready yet and try to wait for it to load. But this would never happen,
alas.

If a scan error happens on opening a new directory, please release the
newly created keydir so other open operations have a chance to succeed.