Fix extra update tombstones
An unrelated fix in commit 4ab96d4 caused updates replacing a value
in the same file to write extra tombstones that serve no purpose.
Updates are supposed to create a tombstone for previous values only
when the previous value was stored in an older file. Multiple
values for the same key in the same file do not require these extra
tombstones.