Add failing update tombstones test
Our current delete scheme requires that updates write tombstones for
previous values when the previous value is in an file logically older
than the current write file.  An unrelated change broke that logic, and
now tombstones are written every single time a value is updated.

This test demonstrates this by counting the number of tombstones
expected for repeatedly updating the same key when two data files are
created. Only a tombstone is required, but we get one per update.