Add failing list data files race test
This test shows what happens when listing the readable files from a
Bitcask directory. The list should not include the active write or merge
files. But if those change after the code fetches the name of the
current one, the listing will contain the new active write and/or merge
files, skipping the *previous* write or merge file. That would be bad.

The test uses meck to return the filename, then in one go write a new
one to simulate the race.