Avoid loading keys from corrupted entries
Due to an optimization added last year, when keys are loaded from a data
file due to a bad or missing hintfile, keys were loaded even if the
record did not pass the CRC test. This was done to avoid extra I/O by
skipping reading the value, but it is incorrect and has caused a few
customers some problems either by Bitcask crashing on startup or data
scans failing later when hitting a corrupt key.