Remove use of prim_file:list_dir/2
The use of it was a hack of some sort - which is not presently understood.  It is no longer valid due to https://github.com/erlang/otp/commit/ebbd26eeea4115c946d1254d94acd50f150b4455 removing the export.

It is assumed that list_dir was originally added to avoid problems no solved via dirty NIF scheduling for file I/O